refactor: collapse app notify thin compat boundaries
